refactor: migrate to modular server actions structure
