+++ app/actions/index.ts
@@ -1,5 +1,6 @@
 import * as users from "./users"
 import * as sessions from "./sessions"
 import * as votes from "./votes"
+import * as utils from "./utils"
 
-export { users, sessions, votes } 
+export { users, sessions, votes, utils }
