+++ app/actions/users.ts
@@ -1,6 +1,6 @@
 "use server"
 
-import { createActionSupabaseClient } from "@/lib/supabase-server"
+import { createActionSupabaseClient } from "@/lib/supabase-actions"
 import { ApiResponse } from "./types"
 import type { User } from "@/types"
 import { handleSupabaseError } from "./utils"
