+++ lib/supabase-actions.ts
@@ -1,0 +1,13 @@
+import { createClient } from "@supabase/supabase-js"
+
+// Create a server action client (for use in Server Actions)
+export async function createActionSupabaseClient() {
+  const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL
+  const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
+  
+  if (!supabaseUrl || !supabaseKey) {
+    throw new Error('Missing Supabase credentials')
+  }
+  
+  return createClient(supabaseUrl, supabaseKey)
+}
